Extract PushPackageAsync into GalleryOperations to remove duplication
