+++ src/GalleryTools/Commands/GalleryOperations.cs
@@ -2,7 +2,11 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.IO;
 using System.Linq;
+using System.Net;
+using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using NuGet.Services.Entities;
 using NuGetGallery;
@@ -119,6 +123,62 @@ namespace GalleryTools.Commands
 
 			Console.WriteLine ($"Created API key '{description}' for '{user.Username}' (scope owner={scopeOwner.Username}).");
 			return plaintextApiKey;
+		}
+
+		/// <summary>
+		/// Pushes a .nupkg file to the Gallery via the HTTP API.
+		/// Handles 409 Conflict (already exists) and 403 Forbidden (owned by another account) gracefully.
+		/// </summary>
+		public static async Task PushPackageAsync(string baseUrl, string apiKey, string nupkgPath)
+		{
+			var fileName = Path.GetFileName(nupkgPath);
+
+			// Trust dev certs for localhost HTTPS
+			var handler = new HttpClientHandler();
+			if (baseUrl.IndexOf("localhost", StringComparison.OrdinalIgnoreCase) >= 0)
+			{
+				handler.ServerCertificateCustomValidationCallback = (msg, cert, chain, errors) => true;
+			}
+
+			using (var client = new HttpClient(handler))
+			{
+				client.DefaultRequestHeaders.Add("X-NuGet-ApiKey", apiKey);
+				client.DefaultRequestHeaders.Add("X-NuGet-Client-Version", "6.0.0");
+				client.Timeout = TimeSpan.FromSeconds(120);
+
+				using (var fileStream = File.OpenRead(nupkgPath))
+				{
+					var content = new MultipartFormDataContent();
+					var streamContent = new StreamContent(fileStream);
+					streamContent.Headers.ContentType = new MediaTypeHeaderValue("application/octet-stream");
+					content.Add(streamContent, "package", fileName);
+
+					var pushUrl = $"{baseUrl.TrimEnd('/')}/api/v2/package";
+					var response = await client.PutAsync(pushUrl, content);
+
+					if (response.StatusCode == HttpStatusCode.Conflict)
+					{
+						Console.WriteLine($"Package {fileName} already exists (409 Conflict). Skipping.");
+						return;
+					}
+
+					if (response.StatusCode == HttpStatusCode.Forbidden)
+					{
+						// Package may already exist under a different owner (e.g. seeded by AppHost).
+						Console.WriteLine($"Package {fileName} returned 403 Forbidden (likely already owned by another account). Skipping.");
+						return;
+					}
+
+					if (!response.IsSuccessStatusCode)
+					{
+						var body = await response.Content.ReadAsStringAsync();
+						throw new InvalidOperationException(
+							$"Failed to push {fileName}: {response.StatusCode} {response.ReasonPhrase}\n{body}");
+					}
+
+					Console.WriteLine($"Pushed {fileName} via API ({response.StatusCode}).");
+				}
+			}
 		}
 	}
 }
